노래추가, 웹빌드용 진동함수 및 구글플레이 주석화, 강제플레이, 하트->광고

+++ STAC/Assets/Scripts/Rotate.cs
@@ -12,7 +12,7 @@ public class Rotate : MonoBehaviour
     public bool isRight = true;
     public bool isTriangle = true;
     public float delay;
-    private int bpm;
+    private float bpm;
     private double currentTime = 0d;
     public int value;
     public ParticleSystem[] rotateParticles;
